+++ src/SpecificOTI.java
@@ -2,8 +2,8 @@
 
 public class SpecificOTI {
 
-	private short Z; // number of source blocks
-	private int N; // number of sub-blocks
-	private short Al; // alignment parameter
+	private long Z; // number of source blocks
+	private long N; // number of sub-blocks
+	private long Al; // alignment parameter
 	
 }
